Lint: curate the rule set, clear the backlog, make ruff a real gate
The lint job reported roughly 1,700 findings on every run and was marked
continue-on-error, so it was permanently red and could never tell anyone when
something new broke. ruff.toml's own header said the goal was to "pass clean
so it can act as a gate" -- it just never implemented that.

Root cause: ruff.toml set `ignore` but never `select`, so the effective rule
set was ~413 rules, including whole families that contradict how this codebase
is deliberately written.

ruff.toml now sets an explicit selection (E, W, F, I, UP, B, RUF) and
documents, inline, every family left out and why:

  BLE001  389 hits. Broad excepts are the architecture here -- drive access,
          audio decode and Rekordbox reads degrade rather than crash a long
          batch, and several helpers document "never raises" as a contract.
  S110    58 hits, same reasoning.
  DTZ     72 hits. Worth doing, but timezone changes touch archive and
          snapshot logic and belong in their own change.
  RUF001/2/3  Ambiguous-unicode, 30 hits -- the em-dashes and arrows in the
          section headers are deliberate typography in comments.
  B905    zip(strict=) changes runtime behaviour at 15 call sites; that is a
          per-site correctness review, not a lint sweep.

Then cleared what remained: ~2,200 fixes, mostly import sorting, unused noqa
directives, and PEP 585/604 annotations, plus hand fixes for the findings
that were real:

  * `Optional[callable]` in importer_database and library_browser used the
    builtin FUNCTION, not the Callable TYPE. Optional[] never evaluated it so
    it worked by accident; rewriting to `callable | None` crashed on import.
    Fixed properly to `Callable | None` with the right import.
  * B023 closures over loop variables in importer_database (drive index) and
    live_bridge/daemon (bound host/port), bound as defaults. Both are called
    synchronously today, but imports run on background threads here.
  * F811 in routes_player: get_connected_volumes was imported from helpers
    and then redefined locally. Removed the dead import.
  * B904 exception chaining, B011 assert False, RUF013 implicit Optional,
    E741 ambiguous `l`, and two F401s that are availability probes rather
    than unused imports (documented as such).

CI: ruff is split from pyright and now gates. Keeping them in one job would
drag ruff back into always-red. pyright stays informational -- its backlog
has not been assessed. ruff is pinned in requirements-dev.txt rather than
installed ad hoc, since it is now a gate.

596 tests pass. `ruff check .` reports "All checks passed!".

+++ tests/test_anlz_reader.py
@@ -15,14 +15,12 @@ import struct
 import sys
 from pathlib import Path
 
-import pytest
-
 REPO_ROOT = Path(__file__).resolve().parent.parent
 for p in (REPO_ROOT, REPO_ROOT / "chop_shop"):
     if str(p) not in sys.path:
         sys.path.insert(0, str(p))
 
-from anlz_reader import parse_anlz_file, read_anlz_set  # noqa: E402
+from anlz_reader import parse_anlz_file, read_anlz_set
 
 
 def _tag(fourcc: bytes, body: bytes, len_header: int = 12) -> bytes:
@@ -46,7 +44,7 @@ def _ppth_tag(path: str) -> bytes:
 def _pqtz_tag(beats: list) -> bytes:
     body = struct.pack(">III", 0, 0, len(beats))
     for beat_no, tempo_bpm, time_ms in beats:
-        body += struct.pack(">HHI", beat_no, int(round(tempo_bpm * 100)), time_ms)
+        body += struct.pack(">HHI", beat_no, round(tempo_bpm * 100), time_ms)
     return _tag(b"PQTZ", body)
 
 
@@ -174,7 +172,7 @@ def test_pqtz_truncates_when_declared_beats_exceed_body(tmp_path):
     # return only the beats that fit and must not raise.
     body = struct.pack(">III", 0, 0, 10)  # declares 10 beats
     for beat_no, tempo_bpm, time_ms in [(1, 124.0, 110), (2, 124.0, 594)]:  # only 2 present
-        body += struct.pack(">HHI", beat_no, int(round(tempo_bpm * 100)), time_ms)
+        body += struct.pack(">HHI", beat_no, round(tempo_bpm * 100), time_ms)
     tags = _tag(b"PQTZ", body)
 
     anlz_path = tmp_path / "ANLZ0000.DAT"
